surya/pyproject.toml
[project]
name = "surya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=10.4.0",
    "requests>=2.32.5",
    "surya-ocr>=0.17.1",
]
